requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
requests==2.31.0

# LLM & LangChain
langchain
langsmith
langchain-groq
langchain-community

# Embeddings & Retrieval
sentence-transformers==5.2.0
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Data processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# Web scraping
beautifulsoup4==4.12.2
dotenv
# Configuration
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
pytest==7.4.3

google-generativeai
httpx